6-CovertChannels/HTTP_GIF_Covert_Channel/07_<name>_F1/requirements.txt
absl-py==1.4.0
anyio==3.7.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.2
attrs==21.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
beniget==0.4.1
bleach==6.0.0
blinker==1.6.2
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
dm-tree==0.1.8
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
entrypoints==0.4
etils==1.3.0
exceptiongroup==1.1.1
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.17.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-tools==1.54.2
gyp==0.1
h11==0.14.0
h5py==3.9.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.12
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
jupyterlab_server==2.23.0
keras==2.12.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==16.0.0
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mistune==2.0.5
ml-dtypes==0.2.0
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.0
olefile==0.46
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.5.3
ply==3.11
prometheus-client==0.17.0
promise==2.3
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.18.0
pycups==2.0.1
pyDH==0.1.6
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-json-logger==2.0.7
python-nmap==0.7.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.10.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
ssh-import-id==5.11
stack-data==0.6.2
sympy==1.9
systemd-python==234
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.3.0
terminado==0.17.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==3.1.0
tinycss2==1.2.1
toarray==0.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.3.6
widgetsnbextension==4.0.7
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
